backend/requirements.txt
flask-cors
flask
sentence-transformers
requests
scikit-learn
faiss-cpu

